feat: add type function
